requirements.txt
fastapi
uvicorn
feedparser
httpx
beautifulsoup4
requests
selenium
pytz
pymongo
python-dotenv
google-generativeai
webdriver-manager
lxml
aiohttp
PyJWT